styling changes to editable meeting modal and discard changes modal
