+++ src/layout/Nav/Nav.scss
@@ -11,7 +11,7 @@
 
 nav {
   align-items: center;
-  background-color: #313335;
+  background-color: #5C6BC0;
   display: flex;
   height: 65px;
   justify-content: space-between;
