+++ src/layout/SideMenu/SideMenu.scss
@@ -1,11 +1,12 @@
 .sidemenu-container {
   background: #ffffff;
-  box-shadow: 3px 3px 10px rgba(0, 0, 0, 0.1);
+  box-shadow: 0px 1px 3px 1px #1a237e26;
   display: flex;
   flex-direction: column;
   padding: calc(93px + 1rem) 16px 32px;
   transition: 0.3s ease-in-out;
   width: 272px;
+  border: 1.5px solid #1a237e26;
 
   .completion-overflow-btn {
     align-self: center;
