+++ src/screens/Calendar/MeetingModal/EditableMeetingModal/DiscardChangesButtons.tsx
@@ -5,9 +5,35 @@ export const DiscardChangesButtons = ({
   handleCloseDiscardChanges,
 }) => {
   return (
-    <div>
-      <Button onClick={handleClose}>Cancel</Button>
-      <Button onClick={handleCloseDiscardChanges}>Discard</Button>
+    <div className='discard-buttons-container'>
+      <Button onClick={handleClose} sx={cancelButtonStyles}>
+        Cancel
+      </Button>
+      <Button onClick={handleCloseDiscardChanges} sx={discardButtonStyles}>
+        Discard
+      </Button>
     </div>
   )
+}
+
+const discardButtonStyles = {
+  backgroundColor: '#fa9413',
+  color: '#1A237E',
+  textTransform: 'none',
+  padding: '8px 20px',
+
+  '&:hover': {
+    backgroundColor: '#fa9413',
+  },
+}
+
+const cancelButtonStyles = {
+  color: '#1A237E',
+  backgroundColor: 'white',
+  textTransform: 'none',
+  padding: '8px 20px',
+
+  '&:hover': {
+    backgroundColor: 'white',
+  },
 }
