+++ src/screens/Calendar/MeetingModal/EditableMeetingModal/DiscardChangesModal.tsx
@@ -2,6 +2,7 @@ import Box from '@mui/material/Box'
 import Button from '@mui/material/Button'
 import Modal from '@mui/material/Modal'
 import { DiscardChangesButtons } from './DiscardChangesButtons'
+import '../styles/DiscardChangesModal.scss'
 
 export const DiscardChangesModal = ({
   handleCloseDiscardChanges,
