+++ src/screens/Calendar/MeetingModal/styles/DiscardChangesModal.scss
@@ -1,0 +1,13 @@
+.MuiBox-root.css-1acyv19  {
+  max-width: 300px;
+  padding: 25px;
+  padding-top: 20px;
+  border-radius: 5px;
+}
+
+
+.discard-buttons-container {
+  margin-top: 30px;
+  justify-content: end;
+  display: flex;
+}
