+++ src/screens/Calendar/MeetingModal/styles/MemberCheckbox.scss
@@ -4,6 +4,8 @@
   justify-content: space-between;
   width: 70%;
   margin-left: -10px;
+
+
  
 
   p {
@@ -15,6 +17,27 @@
   .availability-text {
     p {
       text-align: center;
+      color: #212121;
+    }
+  }
+}
+
+.member-checkbox-unavailable {
+  display: flex;
+  flex-direction: row;
+  justify-content: space-between;
+  width: 70%;
+  margin-left: -10px;
+
+  .MuiTypography-root{
+    color: #212121;
+    opacity: 0.55
+  }
+
+  .availability-text {
+    p {
+      text-align: center;
+      color: #212121;
     }
   }
 }
